feat: divide table view data source from TagManagementViewController
#15
